Fix segment dault issue.

+++ src/engine/widgets/CircleWidget.cpp
@@ -2,7 +2,7 @@
 #include "CircleWidget.h"
 
 
-CircleWidget::CircleWidget(const int& id, b2Body* body,const std::string& tag, b2BodyDef* bodyDefinition, const float& radius): BaseWidget(id, body, tag, bodyDefinition), m_radius(radius){
+CircleWidget::CircleWidget(const int& id, b2Body* body,const std::string& tag, b2BodyDef* bodyDefinition, const double& radius): BaseWidget(id, body, tag, bodyDefinition), m_radius(radius){
     
 }
